requirements.txt
python-telegram-bot==20.7
anthropic>=0.25.0
apscheduler==3.10.4
python-dotenv==1.0.0
openai>=1.0.0        # Whisper API для голосовых сообщений
